Add status label to total requests.
This change moves the total request counter form the `prepare` function to the `on_response` so that we can capture the status code of the response and emit that as a par of the metric.

The reasoning behind this is that with our usage of this lib and axum, we want to be able to graph total service QPS by method, endpoint, as well as status (e.g. has X endpoint had a spike in 500's etc).